refactor: 拆分agent快照窗口事务分发功能
